Temp-Mail/requirements.txt
fastapi
uvicorn
aiohttp
cloudscraper
beautifulsoup4
brotli
zstandard
python-telegram-bot[ext]
python-dotenv
flask